% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_sticker.R
\name{create_sticker}
\alias{create_sticker}
\title{Create a hex sticker for a package.}
\usage{
create_sticker(pkg_name = "delemr")
}
\arguments{
\item{pkg_name}{A string with the package name.}
}
\value{
Nothing, but creates a sticker in the \code{man/figures} folder and opens
it in the viewer.
}
\description{
This is the code I used to create the sticker for the \code{delemr} package.

I created this function to remember some cool tricks I found on the internet
to create the sticker of this package from an image while bypassing the
limitations of the \code{hexSticker} package. This is mostly the use of the
\code{cropcircles} package to crop the image, the \code{showtext} package to use a
custom font, and the creation of a precise ggplot object to have way more
control over the final sticker.

The function is not flexible enough to be exported, although I allowed to
specify a package name. In that setting, putting the raw image in the
\code{man/figures} folder and name it \code{pkg_name_raw_img.png} should suffice to run
the function. However, the ggplot call will a "delemr" text on the sticker,
so this code still has to be tweaked to avoid that and become truly flexible.
}
