% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_contrasts.R
\name{define_contrasts}
\alias{define_contrasts}
\title{Define contrasts inside a pipeline}
\usage{
define_contrasts(df, col, contrast)
}
\arguments{
\item{df}{A dataframe.}

\item{col}{A column name.}

\item{contrast}{A matrix of contrasts.}
}
\value{
A dataframe with the contrasts defined.
}
\description{
Short wrapper around the \code{contrasts} function to define contrasts inside a pipeline easily. The function takes a dataframe as input and returns a dataframe with updated contrasts, allowing to pipe it, which differs from the usual use of the \code{contrasts} function.
}
\examples{
df <- data.frame(group = factor(rep(c("A", "B", "C"), 3)))

# The usual way
contrasts(df$group) <- contr.sum(3)

# the pipeline way
df <- df |> define_contrasts(group, contr.sum(3))

contrasts(df$group)
}
