% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg-themes.R
\name{theme_transparent}
\alias{theme_transparent}
\alias{theme_element_colour}
\title{Theme modifiers for ggplot2}
\usage{
theme_transparent()

theme_element_colour(colour = "black")
}
\arguments{
\item{colour}{The colour to set all small line and text elements to.}
}
\description{
These functions modify specific theme elements of a plot. Be wary that,
contrary the usual "theme" functions (e.g., \code{\link[=theme_bw]{theme_bw()}}), these do not
"reset" all current elements, they only modify some of them.
}
\details{
\describe{

\item{\code{theme_transparent()}}{
Sets the background of the plot, panel, and legend to transparent.}

\item{\code{theme_element_colour()}}{
Sets the colour of all small line and text elements to a specified colour.}
}
}
\examples{
library(ggplot2)

p <- ggplot(mtcars, aes(mpg, wt)) + geom_point()

p + theme_transparent()
p + theme_element_colour("red")

}
\seealso{
\code{\link[=theme]{theme()}}, \code{\link[=theme_bw]{theme_bw()}}
}
