% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{glue_col_lit}
\alias{glue_col_lit}
\title{Use glue syntax with literal interpretation}
\usage{
glue_col_lit(...)
}
\arguments{
\item{...}{Character vectors to be concatenated}
}
\description{
Use glue syntax with literal interpretation
}
