% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{quote_vector}
\alias{quote_vector}
\title{Add quotes to a vector with quoted or unquoted elements}
\usage{
quote_vector(vect)
}
\arguments{
\item{vect}{A vector of elements that can be quoted or unquoted, e.g.,
\code{c(a, b, c)}, \code{c("a", b, "c")}, \code{c(a, "b", c)}, etc.}
}
\value{
A vector with all the elements quoted, i.e., turned into strings.
}
\description{
Add quotes to a vector with quoted or unquoted elements
}
