% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_consecutive.R
\name{sum_consecutive}
\alias{sum_consecutive}
\title{Sum non-null consecutive values in a vector}
\usage{
sum_consecutive(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
A vector where consecutive non-null numeric values have been summed.
}
\description{
Function to sum  non-null consecutive numeric values in a vector containing
zeros. Multiple successive numbers are all replaced with their sum and all
zeros are removed.
}
\examples{
sum_consecutive(c(1, 2, 0, 0, 3, 4, 0, 0, 0, 4, 3))
sum_consecutive(c(1, 2, 0, 12, 1, 3, 0, 0, 8, 0, 3))
}
