% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_consecutive_values.R
\name{sum_consecutive_values}
\alias{sum_consecutive_values}
\title{Sum non-null consecutive values in a vector}
\usage{
sum_consecutive_values(x, unique = FALSE)
}
\arguments{
\item{x}{A vector.}

\item{unique}{Logical. If \code{TRUE}, the function will return only unique values.}
}
\value{
A vector where consecutive non-null numeric values have been summed.
}
\description{
Function to sum consecutive non-null numeric values in a vector containing zeros. Multiple successive numbers are all replaced with their sum (see examples). Duplicates can be removed by setting the \code{unique = TRUE}.
}
\examples{
sum_consecutive_values(c(1, 2, 0, 0, 3, 4, 0, 0, 0, 5, 6))
sum_consecutive_values(c(1, 2, 0, 0, 3, 4, 0, 0, 0, 5, 6), unique = TRUE)
}
