% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delemr-package.R
\docType{package}
\name{delemr-package}
\alias{delemr}
\alias{delemr-package}
\title{delemr: Maël Delem's R code}
\description{
Miscellaneous R functions developed in personal projects.
}
\author{
\strong{Maintainer}: Maël Delem \email{mael.a.delem@gmail.com} (\href{https://orcid.org/0009-0005-8518-1991}{ORCID})

}
\keyword{internal}
